backend/requirements.txt
# FastAPI and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.13.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
cryptography==41.0.7

# Email validation
email-validator==2.1.0

# Environment variables
python-dotenv==1.0.0

# Logging and monitoring
python-json-logger==2.0.7

# Rate limiting
slowapi==0.1.9

# Data processing
numpy==1.26.2
pandas==2.1.3
scikit-learn==1.3.2
nltk==3.8.1
joblib==1.3.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Development
black==23.12.0
flake8==6.1.0
isort==5.13.2
mypy==1.7.1

# Documentation
swagger-ui==3.52.5

# Production server (optional)
gunicorn==21.2.0

# Optional: Database drivers (uncomment as needed)
# psycopg2-binary==2.9.9  # PostgreSQL
# pymysql==1.1.0  # MySQL

# Optional: Speech-to-text (choose one or more for production)
# google-cloud-speech==2.21.0  # Google Cloud
# boto3==1.29.7  # AWS Transcribe
# azure-cognitiveservices-speech==1.34.0  # Azure Speech
# openai==1.3.6  # OpenAI Whisper API
